requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Flask-SQLAlchemy==1.0.0
Jinja2==2.7.2
MarkupSafe==0.23
Werkzeug==0.9.4
wsgiref==0.1.2
raven==4.1.1
blinker==1.3
MySQL-python==1.2.5